+++ YouTube.Uwp/App.xaml.cs
@@ -1,5 +1,7 @@
 using Windows.ApplicationModel;
 using Windows.ApplicationModel.Activation;
+using System;
+using System.Threading.Tasks;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using YouTube.Uwp.Services;
@@ -12,6 +14,8 @@ namespace YouTube.Uwp
         {
             InitializeComponent();
             Suspending += OnSuspending;
+            UnhandledException += OnUnhandledException;
+            TaskScheduler.UnobservedTaskException += OnUnobservedTaskException;
             Configuration = new RuntimeConfiguration();
         }
 
@@ -19,6 +23,7 @@ namespace YouTube.Uwp
 
         protected override void OnLaunched(LaunchActivatedEventArgs e)
         {
+            DiagnosticLog.Write("App.Launch", "Application launched.");
             EnsureMainPage();
             Window.Current.Activate();
         }
@@ -40,8 +45,20 @@ namespace YouTube.Uwp
 
         private void OnSuspending(object sender, SuspendingEventArgs e)
         {
+            DiagnosticLog.Write("App.Suspend", "Application suspending.");
             SuspendingDeferral deferral = e.SuspendingOperation.GetDeferral();
             deferral.Complete();
+        }
+
+        private void OnUnhandledException(object sender, UnhandledExceptionEventArgs e)
+        {
+            DiagnosticLog.WriteException("App.UnhandledException", e.Exception);
+        }
+
+        private void OnUnobservedTaskException(object sender, UnobservedTaskExceptionEventArgs e)
+        {
+            DiagnosticLog.WriteException("App.UnobservedTask", e.Exception);
+            e.SetObserved();
         }
     }
 }